speakeasy-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.0
google-cloud-aiplatform>=1.60.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
